Setup jetty webserver with a Maven pom file
And, added documentation to the README for it + some markup edits

This lets a user quickly get up and running without having to install a
webserver first. Since we're already familiar with Maven, I used that to
set it up.

Keep in mind that you don't have to use maven; the pom file is just a
convenience.